сделать поиск + [Authorize]

+++ Libr/Data/Models/book.cs
@@ -13,7 +13,7 @@ namespace Libr.Models
         public string Name { get; set; }
         public string Author { get; set; }
         public string description { get; set; }
-        public ushort count { get; set; }
+        public int count { get; set; }
         public float price { get; set; }
         public string PhotoPath { get; set; }
         public Genre? genres { get; set; }
